+++ src/main/java/com/vomiter/survivorsdelight/adapter/cooking_pot/PotFluidRecipeHooks.java
@@ -1,0 +1,4 @@
+package com.vomiter.survivorsdelight.adapter.cooking_pot;
+
+public class PotFluidRecipeHooks {
+}