+++ src/main/java/com/vomiter/survivorsdelight/content/device/cooking_pot/fluid_handle/ICookingPotFluidAccess.java
@@ -6,13 +6,15 @@ import net.neoforged.neoforge.fluids.capability.IFluidHandler;
 import net.neoforged.neoforge.items.ItemStackHandler;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.List;
+
 public interface ICookingPotFluidAccess {
     IFluidTank sdtfc$getTank();
     ItemStackHandler sdtfc$getAuxInv();
-    ItemStackHandler sdtfc$getInventory();
     void sdtfc$updateFluidIOSlots();
     void sdtfc$addPlayer(ServerPlayer player);
     void sdtfc$removePlayer(ServerPlayer player);
+    List<ServerPlayer> sdtfc$getPlayer();
 
     default @Nullable IFluidHandler sd$getFluidHandler() {
         IFluidTank tank = sdtfc$getTank();