+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 ## 23rd
 - Refactor: Cabinet
 - Compat: FL; copied from 1.20.1, not tested yet
+- Refactor: Cooking Pot, WIP